+++ Debug/DPrint.c
@@ -18,7 +18,7 @@ void DPrint(const char *str, const uint color)
 {
 #ifdef ENABLE_DEBUG_PRINT
 	FontDrawString((Vector2){12, dprintYPosition + 2}, str, 16, 0xFF000000, true);
-	dprintYPosition += FontDrawString((Vector2){10, dprintYPosition}, str, 16, color, true).y - dprintYPosition + 8;
+	dprintYPosition += (int)FontDrawString((Vector2){10, dprintYPosition}, str, 16, color, true).y - dprintYPosition + 8;
 #endif
 }
